backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
ecdsa==0.19.0
fastapi==0.115.4
filelock==3.16.1
frozenlist==1.5.0
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
identify==2.6.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.6
langchain-community==0.3.5
langchain-core==0.3.15
langchain-groq==0.2.0
langchain-text-splitters==0.3.1
langsmith==0.1.138
Markdown==3.7
markdownify==0.13.1
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==1.26.4
orjson==3.10.10
packaging==24.1
pandas==2.2.3
passlib==1.7.4
platformdirs==4.3.6
propcache==0.2.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyMuPDF==1.24.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.2
tenacity==9.0.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uuid==1.30
uvicorn==0.32.0
virtualenv==20.27.1
yarl==1.17.1
